Struct option_future::OptionFuture

pub struct OptionFuture<F> { /* fields omitted */ }

Effectively a size=1 future queue, or a slot with at-most-one future inside. This struct is mostly useful for modeling an optional path in futures::select!.

For now, we require F: Unpin which mandates boxing of async fn futures but simplifies the OptionFuture implementation.

Methods

impl<F> OptionFuture<F> where
F: Future + Unpin,

pub fn new(inner: Option<F>) -> Self

pub fn or_insert_with<G>(&mut self, fun: G) where
G: FnOnce() -> Option<F>,

If the slot is empty, call the given closure to generate a new future to put into the slot.

Note: the caller is responsible for polling the OptionFuture afterwards.